CUDADistributedTools.jl

A set of utility tools for multi-GPU + multi-process workflows with CUDA.jl + Distributed.jl.

CUDADistributedTools.assign_GPU_workers — Method

assign_GPU_workers(;
    print_info = true, 
    remove_oversubscribed_workers = false
    master_has_own_gpu = false, 
)

Assign each Julia process a unique GPU. Assignment is done using CUDA.device!. Works with workers which may be distributed across different hosts, and with each host possibly having multiple GPUs.

Keyword arguments

remove_oversubscribed_workers — If true, remove worker processes for which there are no free GPUs left. If false, an error is raised instead in this case.

master_has_own_gpu — Controls whether the master process counts as using its GPU, or whether a worker is free to also be assigned the same GPU.

print_info - Print CUDADistributedTools.proc_info after assignment.

Example

julia> using Distributed, CUDA, CUDADistributedTools

julia> CUDA.devices()
CUDA.DeviceIterator() for 4 devices:
0. NVIDIA A100-SXM4-40GB
1. NVIDIA A100-SXM4-40GB
2. NVIDIA A100-SXM4-40GB
3. NVIDIA A100-SXM4-40GB

julia> addprocs(3)
3-element Vector{Int64}:
 2
 3
 4

julia> assign_GPU_workers()
┌ Info: Processes (4):
│  (myid = 1, host = nid001293, device = CuDevice(0): NVIDIA A100-SXM4-40GB 1c40175b))
│  (myid = 2, host = nid001293, device = CuDevice(1): NVIDIA A100-SXM4-40GB f179efe2))
│  (myid = 3, host = nid001293, device = CuDevice(2): NVIDIA A100-SXM4-40GB 36d32866))
└  (myid = 4, host = nid001293, device = CuDevice(3): NVIDIA A100-SXM4-40GB 634451b9))

CUDADistributedTools.proc_info — Method

proc_info()

Prints info about available processes and the GPUs they are assigned.